+++ application/controllers/pengelola/Kelas.php
@@ -24,28 +24,46 @@ class Kelas extends CI_Controller {
 
 	function post()
 	{
-			$this->mod_kelas->save();
-			redirect('pengelola/kelas');
+		if ($this->session->userdata('level') == 'pengelola') {
+				$this->mod_kelas->save();
+				redirect('pengelola/kelas');
+		}
+		else
+		{
+			redirect('home');
+		}
 	}
 
 	function edit()
 	{
-		if (isset($_POST['submit'])) {
-				$this->mod_spesialis->update();
-				redirect('pengelola/kelas');
-		}else {
-				$data['record']= $this->mod_kelas->select_all()->result();
-				$id          = $this->uri->segment(4);
-				$data['row'] = $this->db->get_where('tbl_kelas',array('id_kelas'=>$id))->row_array();
-				$this->load->view('pengelola/edit-kelas',$data);
+		if ($this->session->userdata('level') == 'pengelola') {
+				if (isset($_POST['submit'])) {
+						$this->mod_spesialis->update();
+						redirect('pengelola/kelas');
+				}else {
+						$data['record']= $this->mod_kelas->select_all()->result();
+						$id          = $this->uri->segment(4);
+						$data['row'] = $this->db->get_where('tbl_kelas',array('id_kelas'=>$id))->row_array();
+						$this->load->view('pengelola/edit-kelas',$data);
+				}
+		}
+		else
+		{
+			redirect('home');
 		}
 	}
 
 	function delete()
 	{
-		$this->db->where('id_kelas',$this->uri->segment(4));
-		$this->db->delete('tbl_kelas');
-		redirect('pengelola/kelas');
+		if ($this->session->userdata('level') == 'pengelola') {
+				$this->db->where('id_kelas',$this->uri->segment(4));
+				$this->db->delete('tbl_kelas');
+				redirect('pengelola/kelas');
+		}
+		else
+		{
+			redirect('home');
+		}
 	}
 
 }
