+++ application/controllers/pengelola/User.php
@@ -11,8 +11,8 @@ class User extends CI_Controller {
 
 	public function index()
 	{
-		if ($this->session->userdata('level') == 'pengelola')
 		{
+			if ($this->session->userdata('level') == 'pengelola')
 			$data['record']= $this->mod_user->select_all()->result();
 			$this->load->view('pengelola/user',$data);
 		}
@@ -24,31 +24,49 @@ class User extends CI_Controller {
 
 	function post()
 	{
-		if (isset($_POST['submit'])) {
-				$this->mod_user->save();
-				redirect('pengelola/user');
-		}else {
-			$this->load->view('pengelola/tambah-user');
+		if ($this->session->userdata('level') == 'pengelola') {
+				if (isset($_POST['submit'])) {
+						$this->mod_user->save();
+						redirect('pengelola/user');
+				}else {
+					$this->load->view('pengelola/tambah-user');
+				}
+		}
+		else
+		{
+			redirect('home');
 		}
 	}
 
 	function edit()
 	{
-		if (isset($_POST['submit'])) {
-				$this->mod_user->update();
-				redirect('pengelola/user');
-		}else {
-				$id          = $this->uri->segment(4);
-				$data['row'] = $this->db->get_where('tbl_user',array('id_user'=>$id))->row_array();
-				$this->load->view('pengelola/edit-user',$data);
+		if ($this->session->userdata('level') == 'pengelola') {
+				if (isset($_POST['submit'])) {
+						$this->mod_user->update();
+						redirect('pengelola/user');
+				}else {
+						$id          = $this->uri->segment(4);
+						$data['row'] = $this->db->get_where('tbl_user',array('id_user'=>$id))->row_array();
+						$this->load->view('pengelola/edit-user',$data);
+				}
+		}
+		else
+		{
+			redirect('home');
 		}
 	}
 
 	function delete()
 	{
-		$this->db->where('id_user',$this->uri->segment(4));
-		$this->db->delete('tbl_user');
-		redirect('pengelola/user');
+		if ($this->session->userdata('level') == 'pengelola') {
+				$this->db->where('id_user',$this->uri->segment(4));
+				$this->db->delete('tbl_user');
+				redirect('pengelola/user');
+		}
+		else
+		{
+			redirect('home');
+		}
 	}
 
 }
