+++ application/controllers/perawat/PasienMasuk.php
@@ -13,6 +13,7 @@ class PasienMasuk extends CI_Controller {
 	{
 		if ($this->session->userdata('level') == 'perawat')
 		{
+			$data['ruang']= $this->mod_pasien_masuk->select_ruang()->result();
 			$data['record']= $this->mod_pasien_masuk->select_all()->result();
 			$this->load->view('perawat/pasien-masuk',$data);
 		}
@@ -24,37 +25,69 @@ class PasienMasuk extends CI_Controller {
 
 	function post()
 	{
-			if (isset($_POST['submit'])) {
-					$this->mod_pasien_masuk->save();
-					redirect('perawat/pasienmasuk');
-			}else {
-				$data['ruang']= $this->mod_pasien_masuk->select_ruang()->result();
-				$data['kelas']= $this->mod_pasien_masuk->select_kelas()->result();
-				$data['spesialis']= $this->mod_pasien_masuk->select_spesialis()->result();
-				$this->load->view('perawat/tambah-pmasuk',$data);
-			}
+		if ($this->session->userdata('level') == 'perawat') {
+				if (isset($_POST['submit'])) {
+						$this->mod_pasien_masuk->save();
+						redirect('perawat/pasienmasuk');
+				}else {
+					$data['ruang']= $this->mod_pasien_masuk->select_ruang()->result();
+					$data['kelas']= $this->mod_pasien_masuk->select_kelas()->result();
+					$data['spesialis']= $this->mod_pasien_masuk->select_spesialis()->result();
+					$this->load->view('perawat/tambah-pmasuk',$data);
+				}
+		}
+		else
+		{
+			redirect('home');
+		}
 	}
 
 	function edit()
 	{
-		if (isset($_POST['submit'])) {
-				$this->mod_pasien_masuk->update();
-				redirect('perawat/pasienmasuk');
-		}else {
-				$id          			= $this->uri->segment(4);
-				$data['row'] 			= $this->db->get_where('tbl_pasien_masuk',array('no_rm'=>$id))->row_array();
-				$data['ruang']		= $this->mod_pasien_masuk->select_ruang()->result();
-				$data['kelas']		= $this->mod_pasien_masuk->select_kelas()->result();
-				$data['spesialis']= $this->mod_pasien_masuk->select_spesialis()->result();
-				$this->load->view('perawat/edit-pmasuk',$data);
+		if ($this->session->userdata('level') == 'perawat') {
+				if (isset($_POST['submit'])) {
+						$this->mod_pasien_masuk->update();
+						redirect('perawat/pasienmasuk');
+				}else {
+						$id          			= $this->uri->segment(4);
+						$data['row'] 			= $this->db->get_where('tbl_pasien_masuk',array('no_rm'=>$id))->row_array();
+						$data['ruang']		= $this->mod_pasien_masuk->select_ruang()->result();
+						$data['kelas']		= $this->mod_pasien_masuk->select_kelas()->result();
+						$data['spesialis']= $this->mod_pasien_masuk->select_spesialis()->result();
+						$this->load->view('perawat/edit-pmasuk',$data);
+				}
+		}
+		else
+		{
+			redirect('home');
 		}
 	}
 
 	function delete()
 	{
-		$this->db->where('no_rm',$this->uri->segment(4));
-		$this->db->delete('tbl_pasien_masuk');
-		redirect('perawat/pasienmasuk');
+		if ($this->session->userdata('level') == 'perawat') {
+			$this->db->where('no_rm',$this->uri->segment(4));
+			$this->db->delete('tbl_pasien_masuk');
+			redirect('perawat/pasienmasuk');
+		}
+		else
+		{
+			redirect('home');
+		}
+	}
+
+	function seleksi()
+	{
+		if ($this->session->userdata('level') == 'perawat') {
+				$id          			= $this->uri->segment(4);
+				$data['row'] 			= $this->db->get_where('tbl_pasien_masuk',array('id_ruang'=>$q))->row_array();
+				$data['record']		= $this->mod_pasien_masuk->select()->result();
+				$this->load->view('perawat/pasien-masuk',$data);
+		}
+		else
+		{
+			redirect('home');
+		}
 	}
 
 }
